+++ test/examples.test.js
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest';
-import { splitCli } from './testHelpers.js'
+import { splitCli } from './utils/helpers.js';
 
 //import flaget from 'flaget'; // test local installed package
 import flaget from '../src/index.js'; // test source file for test coverage