+++ test/package/cjs/test.cjs
@@ -1,0 +1,3 @@
+const parseCli = require('flaget');
+
+console.log('>> flaget: ', parseCli().flags);