+++ test/package/esm/test.mjs
@@ -1,0 +1,3 @@
+import parseCli from 'flaget';
+
+console.log('>> flaget: ', parseCli().flags);